+++ frontend/src/components/auth/Signup.js
@@ -1,7 +1,6 @@
-import React, { useContext, useState } from "react";
+import React, { useState } from "react";
 import Form from "react-bootstrap/Form";
 import Button from "react-bootstrap/Button";
-import PropTypes from 'prop-types';
 import './Signup.css';
 import { Link } from "react-router-dom";
 import configData from "../../config.json";
