+++ frontend/src/components/navbar/Navbar.css
@@ -1,0 +1,119 @@
+.navbar {
+  background-color: #060b26;
+  height: 5rem;
+  display: flex;
+  align-items: center;
+}
+
+#hamburger {
+  margin-left: 1.25rem;
+}
+
+#logo {
+  display: flex;
+  width: fit-content;
+  align-items: center;
+  justify-content: center;
+  position: absolute;
+  right: 0;
+  left: 0;
+  margin: auto;
+  flex-direction: row;
+}
+
+.main-buttons {
+  margin-right: 1.5rem;
+}
+
+.main-button {
+  margin: 0.3rem;
+  font-weight: bold;
+}
+
+.hidden {
+  visibility: hidden;
+}
+
+#logo svg {
+  width: 2.5rem;
+  height: 2.5rem;
+}
+
+.menu-bars {
+  /* margin-left: 2rem; */
+  font-size: 2rem;
+  background: none;
+  text-decoration: none;
+  color: white;
+}
+
+.menu-bars:hover {
+  color: white;
+}
+
+.nav-menu {
+  background-color: #060b26;
+  width: 15.625rem;
+  height: 100vh;
+  display: flex;
+  justify-content: center;
+  position: fixed;
+  top: 0;
+  left: -100%;
+  transition: 850ms;
+  z-index: 10;
+  flex-direction: column;
+}
+
+.nav-menu.active {
+  left: 0;
+  transition: 350ms;
+}
+
+.nav-text {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  padding: .5rem 0rem .5rem 0;
+  list-style: none;
+  height: 3.75rem;
+  width: 100%;
+}
+
+.nav-text a {
+  text-decoration: none;
+  color: #f5f5f5;
+  font-size: 1.125rem;
+  width: 95%;
+  height: 100%;
+  display: flex;
+  align-items: center;
+  padding: 0 2rem;
+  border-radius: .25rem;
+}
+
+.nav-text a:hover {
+  background-color: #1a83ff;
+}
+
+.sidebar-top-items {
+  width: 100%;
+  height: 100%;
+  padding: 0; 
+}
+
+.sidebar-bottom-items {
+  width: 100%;
+  align-items: flex-end;
+  display: flex;
+  padding: 0; 
+}
+
+.navbar-toggle {
+  background-color: #060b26;
+  width: 100%;
+  height: 5rem;
+  display: flex;
+  justify-content: flex-start;
+  align-items: center;
+}
